pyproject.toml
[project]
name = "jurl-shortener"
version = "0.1.0"
description = ""
authors = [
    {name = "Johnathan Barbosa",email = "jbarbosa0504@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
